added constraints to top level and refactored where we put logic for checking file extension
