+++ src/spice_rack/_fs_ops/__init__.py
@@ -4,6 +4,7 @@ from spice_rack._fs_ops import (
     _file_systems as file_systems,
     _path_strs as path_strs,
     _exceptions as exceptions,
+    _constraints as constraints
 )
 
 # maintain simplify imports
